requirements.txt
finnhub-python
pyspark
kafka-python
psycopg2-binary
websockets
websocket-client
avro-python3